requirements.txt
python-docx
wordcloud
matplotlib
nltk
